refactor: move tests files
