+++ RSS-CSS-Selectors/src/components/left-side/left-side.test.ts
@@ -1,4 +1,4 @@
-import { delay } from '../components/left-side/left-side';
+import { delay } from './left-side';
 
 test('the data is peanut butter', () => {
     return delay(300).then((data) => {
